+++ tests/integration/video/v1/test_recording.py
@@ -41,6 +41,7 @@ class RecordingTestCase(IntegrationTestCase):
                 "duration": 0,
                 "container_format": "mka",
                 "codec": "OPUS",
+                "track_name": "A name",
                 "grouping_sids": {
                     "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                 },
@@ -108,6 +109,7 @@ class RecordingTestCase(IntegrationTestCase):
                         "duration": 10,
                         "container_format": "mka",
                         "codec": "OPUS",
+                        "track_name": "A name",
                         "grouping_sids": {
                             "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                             "participant_sid": "PAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"