+++ twilio/rest/api/v2010/account/usage/record/yesterday.py
@@ -354,6 +354,7 @@ class YesterdayInstance(InstanceResource):
         SMS_MESSAGES_FEATURES = "sms-messages-features"
         SMS_MESSAGES_FEATURES_SENDERID = "sms-messages-features-senderid"
         SPEECH_RECOGNITION = "speech-recognition"
+        STUDIO_ENGAGEMENTS = "studio-engagements"
         SYNC = "sync"
         SYNC_ACTIONS = "sync-actions"
         SYNC_ENDPOINT_HOURS = "sync-endpoint-hours"