+++ twilio/rest/video/__init__.py
@@ -37,6 +37,13 @@ class Video(Domain):
         return self._v1
 
     @property
+    def compositions(self):
+        """
+        :rtype: twilio.rest.video.v1.composition.CompositionList
+        """
+        return self.v1.compositions
+
+    @property
     def recordings(self):
         """
         :rtype: twilio.rest.video.v1.recording.RecordingList